refactor: 新トップを共通コンポーネント NewTopContent に1ソース化
/ と /top で新トップの描画・データ取得を複製していたのを、共通の
NewTopContent（データ取得+描画を内包、isMemberのみprop）に集約。
各ページは認証/リダイレクト/metadataだけを持つ:
- / : ログイン→/mypageリダイレクト、isMember=false、indexable
- /top: リダイレクトなし、isMember=購読状態、noindex+canonical=/
これで両ページのデザインが常に一致し、今後の変更も1箇所で済む（ドリフト防止）。

検証: tsc緑 / build緑 / dev の / ・/top 両方が新トップを描画
